[feat💡](new screen): new screen to manage some family actions
This screen was built to manage some actions related to families. This UI
is part of the dashboard, that means thatm only admin's have access to this screen. Were they(admin's/admin) can create a new family, register a new member for a determinated family, and delete a family from the 'database' (family's folder).