[feat💡](new screen): manage some actions of families
After the user registrates a new family this screen show up immediatly after, asking the user to choose between register a father, a mother, a son or a daughter to family. Has a feature, we gonna make register one of the introduced options mandatory, that means that the user can not leave this screen before register a father, a son, a daughter or a mother.