[refactor🔄]: created a more simple method to edit all data.
Previously I was creating functions for each operation inside the 'editing family' process. Now i'm creating a single one for the family, where the user will choose if he wants to edit all info related to the family(location,lastname), edit all info related to the father(name,age), the same happens to the mother, and also edit data related to the childrens, but, in this option the user will have to type the name of the son/dauhter he wants to edit. Because a family can have many childrens, but only one father or only one mother or can have both.