[fix💡]: enters the system using the last name and password.
In the last commit, the system had to be open for me to get acess to certain variables, because when the system stops running it reset all variables/attributes, and their values were lost. Now, I could be able to manage that, by reading values from files. I also re-declared all files on the top of the function and modified the scanner reading method from `nextLine()` to `next()` because, `nextLine()` would be only `true` if inside the file had another line, and that wasn't true, because the inside the password file, the password were written in one line only, so there's no other line.